+++ croquest_main/src/bluetooth_host.hpp
@@ -5,7 +5,7 @@
 
 // Begins looking for all bluetooth devices that are advertising with the given host code.
 // quest-guest-xxx-yyyyyy where x is the host code and y is random (users will never see y)
-void                btScanAndConnect(uint8_t host_code);
+void                btScanAndConnect(uint16_t host_code);
 
 // returns a list of all connected devices to the host. dynamically updated.
 std::vector<String> btConnectedDevices();
